b1.2.4

+++ Net_Codeintp_cs/Modules/Group/Commands/Choice/OptIn.cs
@@ -31,33 +31,40 @@ namespace Net_Codeintp_cs.Modules.Group.Commands.Choice
             string s = receiver.MessageChain.GetPlainMessage();
             if (s == "!optin")
             {
-                if (Json.ObjectExistsInArray("optedout", "groups", "groupid", receiver.GroupId))
+                if (Permission.IsGroupAdmin(receiver.GroupId, receiver.Sender.Id))
                 {
-                    Logger.Info($"已令“{receiver.GroupName} ({receiver.GroupId})”重新加入 2kbit Beta 项目！");
-                    Json.DeleteObjectFromArray("optedout", "groups", "groupid", receiver.GroupId);
-                    Update.Do();
-                    try
+                    if (Json.ObjectExistsInArray("optedout", "groups", "groupid", receiver.GroupId))
                     {
-                        await receiver.SendMessageAsync("已重新加入 2kbit Beta 项目（此群将开始收到此机器人的消息）");
+                        Logger.Info($"已令“{receiver.GroupName} ({receiver.GroupId})”重新加入 2kbit Beta 项目！");
+                        Json.DeleteObjectFromArray("optedout", "groups", "groupid", receiver.GroupId);
+                        Update.Do();
+                        try
+                        {
+                            await receiver.SendMessageAsync("已重新加入 2kbit Beta 项目（此群将开始收到此机器人的消息）");
+                        }
+                        catch (Exception e)
+                        {
+                            Logger.Error("群消息发送失败！");
+                            Logger.Debug($"错误信息：\n{e.Message}");
+                        }
                     }
-                    catch (Exception e)
+                    else
                     {
-                        Logger.Error("群消息发送失败！");
-                        Logger.Debug($"错误信息：\n{e.Message}");
+                        Logger.Warning($"未尝试令“{receiver.GroupName} ({receiver.GroupId})”重新加入 2kbit Beta 项目，因为此群已经加入了！");
+                        try
+                        {
+                            await receiver.SendMessageAsync("无法重新加入 2kbit Beta 项目：此群已经加入");
+                        }
+                        catch (Exception e)
+                        {
+                            Logger.Error("群消息发送失败！");
+                            Logger.Debug($"错误信息：\n{e.Message}");
+                        }
                     }
                 }
                 else
                 {
-                    Logger.Warning($"未尝试令“{receiver.GroupName} ({receiver.GroupId})”重新加入 2kbit Beta 项目，因为此群已经加入了！");
-                    try
-                    {
-                        await receiver.SendMessageAsync("无法重新加入 2kbit Beta 项目：此群已经加入");
-                    }
-                    catch (Exception e)
-                    {
-                        Logger.Error("群消息发送失败！");
-                        Logger.Debug($"错误信息：\n{e.Message}");
-                    }
+                    Logger.Warning($"未尝试令“{receiver.GroupName} ({receiver.GroupId})”重新加入 2kbit Beta 项目，因为执行者权限不足！");
                 }
             }
         }
